Break up CloseableBufferedIterator into its own class
